[SNAP-2351] Fixed result mismatch. (#1112)
* [SNAP-2351] Fixed result mismatch.
This issue fixes a problem with row tables and hash joins.
Also, fixes SNAP-2443 due to plan caching of Range based dfs.

Due to plan caching, we don't re-analyze a SQL plan and re-use the same plan if the queries are differed only by literal constants.
However, row tables need a re-evaluation of filters which they can push down to store, every time they execute. For this filter evaluation is in getPartition() method of RowFormatScanRDD.
The problem was getPartition is not called again once the partitions are determined. To circumvent this problem we set "numPartition" variable of RDD to null in CachedDataFrame.

However, this was causing a problem in HashJoin as the build side RDD is not in the dependency graph.
Moreover, the linkPartitionToBucket flag is an analysis time flag which is used in RDD computation time. This was causing the issue of incorrect partition determination.

a) Kept reference of build side RDDs in DelegateRDD. This was used to make the numPartition null.
b) Stored the linkPartitionToBucket flag in CachedDataFrame and used it for subsequent execution.
c) Excluded RangeExec from plan caching.