Fix to SNAP-2577 (#1165)
This is specific to single vm tests where the same vm gets booted as db, then doubles up as lead and
executor. Column buffer table creation gets fired on an embedded connection which gets created as 
part of booting up of db and on that routing to lead node again happens for table creation which 
should not happen. Embedded connection should be picked up only on Executor/Servers. Added that 
check.