Make condition for replicatedTableJoin stricter
- HashJoinExec has a boolean replicatedTableJoin which hints
it to do join using single partition as it would be local join
and build side will be reding from a single partition.
To set this variable, we used to check if all children are
non partitioned recursively
There could be cases where result of NLJ was partitioned and this
caused only single partition of that NLJ result to be consumed by
parent HashJoin.