Update connect_using_jdbc_driver.md
Minor edits. 
@lizygeogy @kneeraj Can we make sure the following is addressed ? I think all setters like setAutocommit are handled properly. But, there are a bunch of other spark, AQP settings that won't get reset, afaik. If so, we need to let the user know.  

Note i left the doc previously  ... <!--- #### (Jags) it is super important for us to document the caveats - if you set a connection property like autocommit(false) other connection requests on this pool will also have this property set. Not sure if this issue exists. If so, critical to document and provide suggestions. Lizy, please verify. --->