[SNAP-2462] enable common-subexpression elimination for ParamLiterals
ParamLiterals are not constants, so common-subexpression elimination phases of optimizer
may fail when values inside them are actually equal.

This adds a new RefParamLiteral that points to a ParamLiteral which is used
when a constant value is found that is same as one found before.
The hashCode mirrors that of original ParamLiteral while equals is reference equality with original.