[SNAP 2634] - providing snappy data specific strategies in IncrementalExecution (#1201)
As part of these changes, we have removed Snappy's DefaultPlanner and instead of passing Snappydata specific strategies as part of `experimentalMethods.extraStrategies` of `SnappySessionState`.

Also, SnappyAggregationStrategy is disabled for stream queries as it is clashing with StatefulAggregationStrategy which is added by Spark as part of `IncrementalExecution`. This is achieved by setting snappydata.sql.hashAggregateSize to -1 in SnappySession's configuration.
This implies that Snappydata aggregation optimization will be turned off for any usage of that particular SnappySession even for non-streaming queries.