Docv1.0.2.1 delta (#1208)
<Other than the master branch, this branch is separately maintained for publishing the corresponding release version of documentation on Read-the-docs>

Added links in the release note corresponding to the items that have supporting documentation.
Incorporated changes to Spark Extension API Guide
Added How-to topic for connecting tableau to SnappyData.
Edited the Setting up OBDC Driver and Tableau chapter and moved the Tableau content to the how-to topic.
Incorporated review comments.
Add description for server-auth-provider.
Added a section on Handling OOM Error with jvm kill agent in the best practices section.
Added security based syntax and examples for Command Line Utilities.
Added syntax and corresponding examples for PUT INTO in column tables and row tables. Added a warning in column tables for the appropriate usage of syntax.