Fix for InternalGemFireError seen with putIntos
Reason was that ColumnDelta.apply could return original value without changing
anything for stats rows, but it was returning the result of getValueRetain call
(which could be a temporary copy) instead of original region value. This could
lead to the temporary value being released completely after writing to Oplog
so that the value put in region will have incorrect reference count leading
to a cascading set of problems. The specific InternalGemFireError issue is due
to the fact that the internal buffer of this released value was changed during
released to an empty one without doing any accounting for the same.

Fix for SNAP-2756. In some rare cases a value might get completely released
before a thread reaches the second sync block in ColumnFormatValue.decompressValue.
For such cases return null as the result value (caller will read from disk if required).

Some cleanups in ColumnFormatEntry after code reviews.