Fixes for issues seen in different putInto tests (#1213)
- switched all intermediate arrays in decoders/encoders to use BufferAllocator so that it can
  be accounted propertly (only in off-heap config for now)
- added proper close methods everywhere for both encoders and decoders for above and invoke
  the close in ColumnTableScan generated code
- force a repartition on the implicit BATCHID column in column update/delete when there is
  no explicit partitioning in the table so that all changes for a batch are together and get
  applied as a single delta rather than spread across all partitions
- use KEY_COLUMNS as partitioning columns if none have been specified expliciitly
- use SnappyCacheTableCommand for putInto intermediate cache to show the plan of cache properly
- moved task interrupt checks to both column and row iterators so that task execution can
  stop quickly; removed the check from ColumnTableScan generated code
- drop temp table when no caching is done
- reduce default spark.memory.fraction to 0.85
- project on only the table columns (and implicit batchId columns) before caching to reduce cache size
- use the cached intermediate result for both insert and update parts of putInto (instead of just former)

Fix for InternalGemFireError seen with putIntos

Reason was that ColumnDelta.apply could return original value without changing
anything for stats rows, but it was returning the result of getValueRetain call
(which could be a temporary copy) instead of original region value. This could
lead to the temporary value being released completely after writing to Oplog
so that the value put in region will have incorrect reference count leading
to a cascading set of problems. The specific InternalGemFireError issue is due
to the fact that the internal buffer of this released value was changed during
released to an empty one without doing any accounting for the same.

Fix for SNAP-2756. In some rare cases a value might get completely released
before a thread reaches the second sync block in ColumnFormatValue.decompressValue.
For such cases return null as the result value (caller will read from disk if required).

Some cleanups in ColumnFormatEntry after code reviews.

Log orignal DDL string (constructed) for column table creates

Test fixes

- fixed a failure in SplitSnappyClusterDUnitTest.testRowTableCreation due to "ANONYMOUS" schema
  created in previous embedded thrift server test
- removed obsolete ResolveAggregationExpressions which is now handled by RefParamLiteral
- moved snappy-core dependency above for kafka tests to work from IDEA due to conflicting
  class KafkaTestUtils getting picked from spark instead
- fixing failing test: a column table must have at-least one extra column apart from partition column to
  perform put into. With recent code changes, we are making key columns as partition columns if
  partition columns are not explicitly specified. Hence the test started failing as it was having only one
  column and the same column is declared as key column as key columns are now treated as partition
  columns.