Branch 1.0.2.2 (#1231)
* Fix for JDBC driver Jar running with Spark 2.3 + versions. ( java.lang.NoSuchFieldError: MAX_ROUNDED_ARRAY_LENGTH )
The Snappy JDBC driver jar contains a class org.apache.spark.unsafe.array.ByteArrayMethods which has been changed in Spark 2.3.1. When Snappy JDBC driver jar dependency is used, this class is loaded from the JDBC jar instead of Spark 2.3.1 jars  (or later Spark version) and the program errors out as it expects a newer version of the class.

We fixed it by relocating the org.apache.spark.unsafe to io.snappydata.org.apache.spark.unsafe in the build.gradle for the JDBC module.