SNAP-2902 Mismatch in the expected and actual inserted rows
Issue is not insert specific rather is in RemoteEntriesIterator where
statsRows iterator does a fetchNextBatch in its next method (around line
130). In addition to fetching the next batch this causes the current
batch values to be freed. So the higher level iterator
ColumnBatchIterator looks at the last value but finds it to be empty
buffer so skips it. As a result iteration over remote entries
consistently misses one ColumnBatch per 1000 (which is the size of a
batch of keys fetched in RemoteEntriesIterator).