- Added new implementation of Analyzer namely "SnappyAnalyzer" which contains a separate copy of rule batches available in Analyzer
This change was required because keeping a dummy instance of analyzer
only to fetch the list of rule batches in order to inject
StringPromotionCheckForUpdate rule was leading to issues as some of the
rules end up using the dummy analyzer instance internally.

Also it is not possible to access the batches from super reference as
batches is declared as a lazy val in Analyzer.

- Fixing failing test

Allowing assignment of tighter numeric types to decimal type

- Enhanced SnappyTestRunner's failure reporting strategy to provide the
complete failure result instead of only indicating that there was "some"
failure