PIVOT support in SQL parser (#1395)
The SnappyParser support deviates from Spark 2.4 support in that it only allows literals in the
value IN list rather than named expressions. This is a limitation of the pivot() API in pre-Spark 2.4
releases itself. On the contrary, it supports explicit GROUP BY columns with PIVOT instead of
always doing implicit detection (which maybe different from what user needs in some cases).

Both the examples below are identical. Spark 2.4 does not support second variant:

```sql
  select * from (
    select year(day) year, month(day) month, temp
    from dayAvgTemp
  )
  PIVOT (
    CAST(avg(temp) AS DECIMAL(5, 2))
    FOR month IN (1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12)
  )
  ORDER BY year DESC
```

Compared to blog example the IN clause only supports constants and not aliases.

```sql
  select * from (
    select year(day) year, month(day) month, temp
    from dayAvgTemp
  )
  PIVOT (
    CAST(avg(temp) AS DECIMAL(5, 2))
    FOR month IN (1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12)
  )
  GROUP BY year
  ORDER BY year DESC
```

## Changes proposed in this pull request

- parser support for PIVOT clause that resolves to Spark's Pivot() LogicalPlan
- parser also injects group by expressions in the Pivot() plan from explicit GROUP BY clause
- added PIVOT as a non-reserved keyword to the grammar
- removed YEAR, MONTH etc from keyword list so can be freely used
- a simple unit test with an accompanying data file