Snap 3135 (#1424)
This is an experimental utility to execute a given command on selected members of the cluster. The script relies on the entries you specify in locators, servers and leads files in conf directory to identify the members of the cluster. "-copyconf" This is a shortcut command which when specified copies log4j.properties, snappy-env.sh and spark-env.sh configuration files from local machine to all the members.