SNAP-3247; Disable/skip table count on UI
* Add property to enable table count on UI in recovery mode
* Show count as -1, when error while in table scan

Ui will appear as is Recovery Mode, expect the table counts.To see table counts user need to set the property
snappydata.recovery.enableTableCountInUI to true. By default the property is set to false and the table count will be shown as -1.
When the table counts is enabled, if there is an error while reading a particular table, it's count will be -1.