suggested fix for SDENT-171 (#1541)
The purpose of the function collectProjectsAndFilters is to flatten the intermediate Project Nodes.
for eg. if the the tree is project1 -> filter1 -> project2 -> filter2 -> project-3 -> BaseTable
and if the aim is to have a flattened structure like
project1 -> filters -> base table.
and given that final projected columns in project1 ( the output) may be less than project2 , project3, the filter2, may be refering to an attribute which is no longer present in project1. This is resoleved in the function by substituting attribute reference in filter with the Expression present in aliases map with the guarantee that expression in aliases map is for sure part of final projection.
The fix is that while identifying substitutable fields, we only use those fields which are deterministic. And if a project node contains an indeterministic field, we ensure that a new ProjectNode containing that indeterministic field stays, so that a filter above it gets satisfied.